refactor(settings): rename settings container
Fix HTML DOM error.
