+++ src/components/dashboard/Home.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Timer from './timer/Timer';
 import Todo from './todo/Todo';
 import Grid from '@material-ui/core/Grid';
-import SettingsModal from 'containers/Settings';
+import SettingsModal from 'containers/SettingsModal';
 import './Home.css';
 
 class Home extends React.Component {
